make dns resolver usage optional
When you are not using the DNS_RESOLVER env var we will use the default
behaviour of nginx connecting to the application.

When DNS_RESOLVER is set to auto the resolving will be done by the
internally set dns resolver, be it the docker one or another overridden
one.

You can also set DNS_RESOLVER to a fixed ip to resolve the other
containers name.

Signed-off-by: BlackEagle <ike.devolder@gmail.com>